crazy pair programming to include detection, preprocessing, surface contact detection, etc

+++ src/user_calibration.cpp
@@ -60,44 +60,83 @@ void UserCalibration::doMouseCallback(int event, int x, int y, int flags)
     }
 }
 
-void UserCalibration::getBallHSVRange(const cv::Mat &image, cv::Scalar &lower_range, cv::Scalar &upper_range)
+void UserCalibration::getBallHSVRange(const cv::Mat &image, cv::Scalar &lower_range, cv::Scalar &upper_range, bool debug)
 {
     /** waitKey needs to be changed so it waits for a mouse event
      * We also need some type of status window that gives instructions instead of std::cout */
     cv::namedWindow("Ball Calibration", cv::WINDOW_AUTOSIZE);
     cv::setMouseCallback("Ball Calibration", mouseCallback, this);
-    cv::imshow("Ball Calibration", image);
-
-    std::cout << "Click center of the ball" << std::endl;
-    cv::waitKey(0);
-    std::cout << "Click any edge of the ball" << std::endl;
-    cv::waitKey(0);
-
-    cv::destroyWindow("Ball Calibration");
-
-    /** Maybe loop here asking the user whether they are satisfied with the circle */
-    cv::Mat display_image;
-    image.copyTo(display_image);
-    cv::circle(display_image, ball_center, ball_radius, cv::Scalar(0,0,255));
-    cv::imshow("Circle", display_image);
-    cv::waitKey(0);
-
-    cv::destroyWindow("Circle");
-
+	cv::imshow("Ball Calibration", image);
+	
+	bool satisfied = false;
+	
+	std::cout << "Starting ball calibration" << std::endl;
+	
+	do
+	{
+		std::cout << "Click center of the ball" << std::endl;
+		cv::waitKey(0);
+		std::cout << "Click any edge of the ball" << std::endl;
+		cv::waitKey(0);
+		
+		//drawing circle around the clicked ball
+		/** Maybe loop here asking the user whether they are satisfied with the circle */
+		cv::Mat display_image;
+		image.copyTo(display_image);
+		cv::circle(display_image, ball_center, ball_radius, cv::Scalar(0,0,255));
+		cv::imshow("Circle", display_image);
+		cv::waitKey(0);
+		cv::destroyWindow("Circle");
+		
+		//checkin with user if he is satisfied with the clicked region
+		char answer = 'n';
+		std::cout << "Is the circle fully inside the ball? (n/y) : " << std::endl;
+		try
+		{
+			std::cin >> answer;
+			std::cout << "Answer was : " << answer << std::endl;
+		}
+		catch(...)
+		{
+			std::cout << "Error while reading answer" << std::endl;
+		}
+		
+		if(answer == 'y')
+		{
+			satisfied = true;
+		}
+		else
+		{
+			satisfied = false;
+			set_center = false;
+			set_radius = false;
+		}
+		
+	}
+	while(!satisfied);
+	
+	cv::destroyWindow("Ball Calibration");
     cv::Mat empty_image = cv::Mat::zeros(image.rows, image.cols, CV_8UC1);
     cv::circle(empty_image, ball_center, ball_radius, cv::Scalar(255,255,255), -1);
     cv::Mat region_of_interest;
     cv::bitwise_and(image, image, region_of_interest, empty_image);
-    cv::imshow("roi", region_of_interest);
-    cv::waitKey(0);
-    cv::destroyWindow("roi");
+    
+	if(debug)
+	{
+		cv::imshow("roi", region_of_interest);
+		cv::waitKey(0);
+		cv::destroyWindow("roi");
+	}
     
     cv::Mat hsv_image;
     cv::cvtColor(region_of_interest, hsv_image, CV_BGR2HSV);
-    cv::imshow("hsvroi", hsv_image);
-    cv::waitKey(0);
-    cv::destroyWindow("hsvroi");
-
+    
+	if(debug)
+	{
+		cv::imshow("hsvroi", hsv_image);
+		cv::waitKey(0);
+		cv::destroyWindow("hsvroi");
+	}
 
     int min_h = 256;
     int min_s = 256;
@@ -130,19 +169,39 @@ void UserCalibration::getBallHSVRange(const cv::Mat &image, cv::Scalar &lower_ra
     std::cout << "Range: " << lower_range <<", " << upper_range << std::endl;
 }
 
+int UserCalibration::getBallFrameEstimate()
+{
+	int frame_number = 0;
+	
+	std::cout << "Please provide with a frame estimate (number) in which the ball appears in the FOV of the camera" << std::endl;
+	try
+	{
+		std::cin >> frame_number;
+		std::cout << "Skipping to frame number : " << frame_number << std::endl;
+		return frame_number;
+	}
+	catch(...)
+	{
+		std::cout << "Error while reading input integer" << std::endl;
+		return 0;
+	}
+}
+
 void UserCalibration::getLineLimits(const cv::Mat &image, std::vector<cv::Point2i> &line_corner_points)
 {
     cv::namedWindow("Line Calibration", cv::WINDOW_AUTOSIZE);
     cv::setMouseCallback("Line Calibration", mouseCallback, this);
     cv::imshow("Line Calibration", image);
 
-    std::cout << "Click bottom left corner of line" << std::endl;
+	std::cout << "Starting line calibration" << std::endl;
+	
+    std::cout << "Click bottom left corner of line, after that press a key" << std::endl;
     cv::waitKey(0);
-    std::cout << "Click top left corner of line" << std::endl;
+    std::cout << "Click top left corner of line, after that press a key" << std::endl;
     cv::waitKey(0);
-    std::cout << "Click top right corner of line" << std::endl;
+    std::cout << "Click top right corner of line, after that press a key" << std::endl;
     cv::waitKey(0);
-    std::cout << "Click bottom right corner of line" << std::endl;
+    std::cout << "Click bottom right corner of line, after that press a key" << std::endl;
     cv::waitKey(0);
     cv::destroyWindow("Line Calibration");
